+++ tests/TestCase/Reflection/Entity/ClassEntityTest.php
@@ -84,7 +84,11 @@ class ClassEntityTest extends TestCase
      */
     public function testGetConstants()
     {
-        $this->assertContainsOnlyInstancesOf(ConstantEntity::class, $this->Class->getConstants());
+        $constants = $this->Class->getConstants();
+        $this->assertContainsOnlyInstancesOf(ConstantEntity::class, $constants);
+        $this->assertSame(['GENUS', 'LEGS'], $constants->map(function (ConstantEntity $constant) {
+            return $constant->getName();
+        })->toList());
     }
 
     /**
@@ -93,8 +97,6 @@ class ClassEntityTest extends TestCase
      */
     public function testGetInterfaces()
     {
-        $this->assertTrue($this->Class->getInterfaces()->isEmpty());
-
         $interfaces = ClassEntity::createFromName(Car::class)->getInterfaces();
         $this->assertContainsOnlyInstancesOf(ClassEntity::class, $interfaces);
         $this->assertSame([Brake::class, MotorVehicle::class], $interfaces->map(function (ClassEntity $interface) {
@@ -128,7 +130,11 @@ class ClassEntityTest extends TestCase
      */
     public function testGetMethods()
     {
-        $this->assertContainsOnlyInstancesOf(MethodEntity::class, $this->Class->getMethods());
+        $methods = $this->Class->getMethods();
+        $this->assertContainsOnlyInstancesOf(MethodEntity::class, $methods);
+        $this->assertSame(['createPuppy', 'setPuppy', 'doMeow', 'setDescription', 'getType'], $methods->map(function (MethodEntity $method) {
+            return $method->getName();
+        })->toList());
     }
 
     /**
@@ -166,7 +172,11 @@ class ClassEntityTest extends TestCase
      */
     public function testGetProperties()
     {
-        $this->assertContainsOnlyInstancesOf(PropertyEntity::class, $this->Class->getProperties());
+        $properties = $this->Class->getProperties();
+        $this->assertContainsOnlyInstancesOf(PropertyEntity::class, $properties);
+        $this->assertSame(['Puppy', 'description', 'isCat'], $properties->map(function (PropertyEntity $property) {
+            return $property->getName();
+        })->toList());
     }
 
     /**
@@ -184,8 +194,6 @@ class ClassEntityTest extends TestCase
      */
     public function testGetTraits()
     {
-        $this->assertTrue(ClassEntity::createFromName(Car::class)->getTraits()->isEmpty());
-
         $traits = $this->Class->getTraits();
         $this->assertContainsOnlyInstancesOf(ClassEntity::class, $traits);
         $this->assertSame(['ColorsTrait', 'PositionTrait'], $traits->map(function (ClassEntity $trait) {